Improved the RegEx replacement logic.
Now with an enabled or disabled key so that you do not need to delete your code when testing. 
Now with an ID so that future developers can identify the rules quickly. 
Now with a comment key so that the RegEx is humanly readable. 
Also fixed a issue when the match group is empty. It now does not output the name of the matched group.